Migrated to Android x and using coroutines. presentation tests fixed and migrated to android x all tests working. also working via command line. removed rx room added Refactored repo and added network listener channels working snackbar added adding dao tests Room db tests added and code clean up read me updated Flavours added for testing

+++ domain/src/main/java/cleanarcpro/brightowusu/com/cleanarcproj/domain/models/DomainUser.kt
@@ -1,6 +1,6 @@
 package cleanarcpro.brightowusu.com.cleanarcproj.domain.models
 
-data class DomainUser(val userId: Int,
+data class DomainUser(val userId: Long,
                       val name: String,
                       val email: String,
                       val phone: String)
