+++ presentation/src/main/java/cleanarcpro/brightowusu/com/cleanarcproj/models/UIUser.kt
@@ -1,6 +1,6 @@
 package cleanarcpro.brightowusu.com.cleanarcproj.models
 
-data class UIUser(val userId: Int,
+data class UIUser(val userId: Long,
                   val name: String,
                   val email: String,
                   val phone: String)
